backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# MongoDB
mongoengine==0.29.1
pymongo==4.6.1
dnspython==2.4.2

# PostGIS - requires manual setup, skip for Windows development
# django-geo-spaas==0.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
firebase-admin==6.5.0

# Background tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0

# Storage
boto3==1.34.0
django-storages==1.14.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
requests==2.31.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Development
ipython==8.18.1
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Geo
django-rest-framework-gis==1.0

